refactor: move the CuTraced Types to Reactant
